// web/src/api/models/NatSpec.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NatKind } from './NatKind';
import {
    NatKindFromJSON,
    NatKindFromJSONTyped,
    NatKindToJSON,
    NatKindToJSONTyped,
} from './NatKind';

/**
 * 
 * @export
 * @interface NatSpec
 */
export interface NatSpec {
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    comment?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    dport?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    family?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    in?: string;
    /**
     * 
     * @type {NatKind}
     * @memberof NatSpec
     */
    kind: NatKind;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    out?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    proto?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    source?: string;
    /**
     * 
     * @type {string}
     * @memberof NatSpec
     */
    to?: string;
}



/**
 * Check if a given object implements the NatSpec interface.
 */
export function instanceOfNatSpec(value: object): value is NatSpec {
    if (!('kind' in value) || value['kind'] === undefined) return false;
    return true;
}

export function NatSpecFromJSON(json: any): NatSpec {
    return NatSpecFromJSONTyped(json, false);
}

export function NatSpecFromJSONTyped(json: any, ignoreDiscriminator: boolean): NatSpec {
    if (json == null) {
        return json;
    }
    return {
        
        'comment': json['comment'] == null ? undefined : json['comment'],
        'dport': json['dport'] == null ? undefined : json['dport'],
        'family': json['family'] == null ? undefined : json['family'],
        'in': json['in'] == null ? undefined : json['in'],
        'kind': NatKindFromJSON(json['kind']),
        'out': json['out'] == null ? undefined : json['out'],
        'proto': json['proto'] == null ? undefined : json['proto'],
        'source': json['source'] == null ? undefined : json['source'],
        'to': json['to'] == null ? undefined : json['to'],
    };
}

export function NatSpecToJSON(json: any): NatSpec {
    return NatSpecToJSONTyped(json, false);
}

export function NatSpecToJSONTyped(value?: NatSpec | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'comment': value['comment'],
        'dport': value['dport'],
        'family': value['family'],
        'in': value['in'],
        'kind': NatKindToJSON(value['kind']),
        'out': value['out'],
        'proto': value['proto'],
        'source': value['source'],
        'to': value['to'],
    };
}