// web/src/api/models/NatKind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const NatKind = {
    KindMasquerade: 'masquerade',
    KindSNAT: 'snat',
    KindDNAT: 'dnat'
} as const;
export type NatKind = typeof NatKind[keyof typeof NatKind];


export function instanceOfNatKind(value: any): boolean {
    for (const key in NatKind) {
        if (Object.prototype.hasOwnProperty.call(NatKind, key)) {
            if (NatKind[key as keyof typeof NatKind] === value) {
                return true;
            }
        }
    }
    return false;
}

export function NatKindFromJSON(json: any): NatKind {
    return NatKindFromJSONTyped(json, false);
}

export function NatKindFromJSONTyped(json: any, ignoreDiscriminator: boolean): NatKind {
    return json as NatKind;
}

export function NatKindToJSON(value?: NatKind | null): any {
    return value as any;
}

export function NatKindToJSONTyped(value: any, ignoreDiscriminator: boolean): NatKind {
    return value as NatKind;
}